extrator/requirements.txt
google-generativeai==0.8.4
python-dotenv==1.0.1
Pillow==10.4.0
